static-rietveld/requirements.txt
google-cloud-storage
bs4
requests
